moved global settings file
